Attempting to reduce jumping effect when new node is created.
Some logic to position new nodes close to their parents instead
of starting all new nodes at a random position. To use, set the
'parent' data property when creating a node to the node's parent.
If this is done, the parent's position plus a random vector will be
used for the new node.